secure-chat-app/client/requirements.txt
PyQt5==5.15.10
requests==2.31.0
websocket-client==1.7.0
cryptography==42.0.0
python-dotenv==1.0.0
